Move debug features into separate components
