+++ app/src/UI/Features/Records/Activity/forms/debug/DebugButton.tsx
@@ -1,0 +1,16 @@
+import { BugReport } from '@mui/icons-material';
+import { Debug } from 'UI/Reusable/Predicates/Debug';
+
+const DebugButton = ({ onClick, label }) => {
+  return (
+    <div className="control">
+      <Debug>
+        <button onClick={onClick}>
+          <BugReport /> {label}
+        </button>
+      </Debug>
+    </div>
+  );
+};
+
+export default DebugButton;
